base app

+++ transaction/transaction_test.go
@@ -2,6 +2,7 @@ package transaction
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"os"
 	"sync"
@@ -17,15 +18,17 @@ func fileExists(filename string) bool {
 }
 
 func TestCreateTransactor(t *testing.T) {
-	const filename = "test_kv"
-	defer os.Remove(filename)
-
 	ctx := context.Background()
-	tr, err := NewFileTransactor(ctx, filename)
+	tr, err := NewFileTransactor(ctx)
 	if err != nil {
 		t.Fatalf("cannot create transactor: %v", err)
 	}
-	defer tr.Close()
+	defer func() {
+		tr.Close()
+		if tr.file != nil {
+			os.Remove(tr.file.Name())
+		}
+	}()
 
 	if tr == nil {
 		t.Fatal("transactor is nil")
@@ -38,7 +41,7 @@ func TestCreateTransactor(t *testing.T) {
 func TestConcurrentWritesAndRead(t *testing.T) {
 	ctx := context.Background()
 
-	transactor, err := NewFileTransactor(ctx, filename)
+	transactor, err := NewFileTransactor(ctx)
 	if err != nil {
 		t.Fatalf("failed to create transactor: %v", err)
 	}
@@ -47,7 +50,9 @@ func TestConcurrentWritesAndRead(t *testing.T) {
 		if err := transactor.Close(); err != nil {
 			t.Fatalf("close error: %v", err)
 		}
-		os.Remove(transactor.file.Name())
+		if transactor.file != nil {
+			os.Remove(transactor.file.Name())
+		}
 	}()
 
 	wg := &sync.WaitGroup{}
@@ -68,7 +73,7 @@ func TestConcurrentWritesAndRead(t *testing.T) {
 	}
 	wg.Wait()
 
-	transactor1, err := NewFileTransactor(ctx, filename)
+	transactor1, err := NewFileTransactor(ctx)
 	if err != nil {
 		t.Fatalf("failed to create transactor1: %v", err)
 	}
@@ -77,11 +82,13 @@ func TestConcurrentWritesAndRead(t *testing.T) {
 		if err := transactor1.Close(); err != nil {
 			t.Fatalf("close error: %v", err)
 		}
-		os.Remove(transactor.file.Name())
+		if transactor.file != nil {
+			os.Remove(transactor.file.Name())
+		}
 	}()
 
 	var readEvents []Event
-	eventsCh, errCh := transactor1.readEvents()
+	eventsCh, errCh := transactor1.ReadEvents()
 	for event := range eventsCh {
 		readEvents = append(readEvents, event)
 	}
@@ -91,5 +98,52 @@ func TestConcurrentWritesAndRead(t *testing.T) {
 	}
 	if len(readEvents) != workers {
 		t.Fatalf("expected %d rows, got %d", workers, len(readEvents))
+	}
+}
+
+func TestSendClosedTransactor(t *testing.T) {
+	ctx := context.Background()
+	tr, err := NewFileTransactor(ctx)
+	if err != nil {
+		t.Fatalf("cannot create transactor: %v", err)
+	}
+	tr.Close()
+
+	defer func() {
+		if tr.file != nil {
+			os.Remove(tr.file.Name())
+		}
+	}()
+
+	const key = "key"
+	const value = "value"
+
+	if err := tr.WritePut(ctx, key, value); !errors.Is(err, ErrTransactorClosed) {
+		t.Fatal("transactor is not closed")
+	}
+
+	if err := tr.WriteDelete(ctx, key); !errors.Is(err, ErrTransactorClosed) {
+		t.Fatal("transactor is not closed")
+	}
+}
+
+func TestClosedTransactor(t *testing.T) {
+	ctx := context.Background()
+	tr, err := NewFileTransactor(ctx)
+	if err != nil {
+		t.Fatalf("cannot create transactor: %v", err)
+	}
+	defer func() {
+		if tr.file != nil {
+			os.Remove(tr.file.Name())
+		}
+	}()
+
+	if err := tr.Close(); err != nil {
+		t.Fatalf("close failed: %v", err)
+	}
+
+	if err := tr.Close(); err != ErrTransactorClosed {
+		t.Fatalf("got: %v, but expected: %v", err, ErrTransactorClosed)
 	}
 }
